Move DistributedDatabase registration out of constructor.
ODistributedDatabaseImpl construction registered the instance, leaking the this reference, and shut down the previous instance if present.
The previous instance may not have been constructed fully however, so shutdown could NPE, resulting in the construction of the current instance aborting with uninitialised state, which would then be picked up by other threads finding it registered in the message service.

This change externalises the construction into an atomic operation in the message service, and makes the state in the distributed database impl final.

The warning about needing registration because of use "further in the call chain" appears to be spurious.