Correct records size after update and delete.
Correctly calculate records size for a cluster on update and delete, adjusting for cluster and chunk entry overheads, and allowing multi-chunk records.
Add checks for record size to all cluster ITs, implemented for V2 and left as a TODO for V0/V1 (which are either incorrect or have a different interpretation of records size).